Crate epic_chain

The block chain itself, validates and accepts new blocks, handles reorgs.

Re-exports

pub use crate::store::ChainStore;

pub use crate::types::BlockStatus;

pub use crate::types::ChainAdapter;

pub use crate::types::Options;

pub use crate::types::SyncState;

pub use crate::types::SyncStatus;

pub use crate::types::Tip;

pub use crate::types::TxHashsetWriteStatus;

Modules

pipe

Implementation of the chain block acceptance (or refusal) pipeline.

store

Implements storage primitives required by the chain

txhashset

Utility structs to handle the 3 hashtrees (output, range proof, kernel) more conveniently and transactionally.

types

Base types that the block chain pipeline requires.

Structs

Chain

Facade to the blockchain block processing pipeline and storage. Provides the current view of the TxHashSet according to the chain state. Also maintains locking for the pipeline to avoid conflicting processing.

Error

Error definition

Enums

ErrorKind

Chain error definitions

Constants

MAX_ORPHAN_SIZE

Orphan pool size is limited by MAX_ORPHAN_SIZE