Enum epic_chain::ErrorKind

pub enum ErrorKind {

    Unfit(String),
    Orphan,
    DifficultyTooLow,
    DifficultyTooHigh,
    WrongTotalDifficulty,
    LowEdgebits,
    InvalidScaling,
    InvalidPow,
    OldBlock,
    InvalidBlockProof(Error),
    InvalidBlockTime,
    InvalidBlockHeight,
    InvalidRoot,
    InvalidMMRSize,
    Keychain(Error),
    Secp(Error),
    AlreadySpent(Commitment),
    DuplicateCommitment(Commitment),
    ImmatureCoinbase,
    MerkleProof,
    OutputNotFound,
    RangeproofNotFound,
    TxKernelNotFound,
    OutputSpent,
    InvalidBlockVersion(HeaderVersion),
    InvalidTxHashSet(String),
    StoreErr(Error, String),
    FileReadErr(String),
    SerErr(Error),
    TxHashSetErr(String),
    TxLockHeight,
    GenesisBlockRequired,
    Transaction(Error),
    Other(String),
    Committed(Error),
    Stopped,
    Bitmap,
    SyncError(String),
    InvalidSortAlgo,
    ThereIsNotPolicy,
    PolicyIsNotAllowed,
    InvalidSeed,

}

Chain error definitions

Variants

Unfit(String)

The block doesn’t fit anywhere in our chain

Tuple Fields of Unfit

0: String

Orphan

Special case of orphan blocks

DifficultyTooLow

Difficulty is too low either compared to ours or the block PoW hash

DifficultyTooHigh

WrongTotalDifficulty

Addition of difficulties on all previous block is wrong

LowEdgebits

Block header edge_bits is lower than our min

InvalidScaling

Scaling factor between primary and secondary PoW is invalid

InvalidPow

The proof of work is invalid

OldBlock

Peer abusively sending us an old block we already have

InvalidBlockProof(Error)

The block doesn’t sum correctly or a tx signature is invalid

Tuple Fields of InvalidBlockProof

0: Error

InvalidBlockTime

Block time is too old

InvalidBlockHeight

Block height is invalid (not previous + 1)

InvalidRoot

One of the root hashes in the block is invalid

InvalidMMRSize

One of the MMR sizes in the block header is invalid

Keychain(Error)

Error from underlying keychain impl

Tuple Fields of Keychain

0: Error

Secp(Error)

Error from underlying secp lib

Tuple Fields of Secp

0: Error

AlreadySpent(Commitment)

One of the inputs in the block has already been spent

Tuple Fields of AlreadySpent

0: Commitment

DuplicateCommitment(Commitment)

An output with that commitment already exists (should be unique)

Tuple Fields of DuplicateCommitment

0: Commitment

ImmatureCoinbase

Attempt to spend a coinbase output before it sufficiently matures.

MerkleProof

Error validating a Merkle proof (coinbase output)

OutputNotFound

Output not found

RangeproofNotFound

Rangeproof not found

TxKernelNotFound

Tx kernel not found

OutputSpent

output spent

InvalidBlockVersion(HeaderVersion)

Invalid block version, either a mistake or outdated software

Tuple Fields of InvalidBlockVersion

0: HeaderVersion

InvalidTxHashSet(String)

We’ve been provided a bad txhashset

Tuple Fields of InvalidTxHashSet

0: String

StoreErr(Error, String)

Internal issue when trying to save or load data from store

Tuple Fields of StoreErr

0: Error1: String

FileReadErr(String)

Internal issue when trying to save or load data from append only files

Tuple Fields of FileReadErr

0: String

SerErr(Error)

Error serializing or deserializing a type

Tuple Fields of SerErr

0: Error

TxHashSetErr(String)

Error with the txhashset

Tuple Fields of TxHashSetErr

0: String

TxLockHeight

Tx not valid based on lock_height.

GenesisBlockRequired

No chain exists and genesis block is required

Transaction(Error)

Error from underlying tx handling

Tuple Fields of Transaction

0: Error

Other(String)

Anything else

Tuple Fields of Other

0: String

Committed(Error)

Error from summing and verifying kernel sums via committed trait.

Tuple Fields of Committed

0: Error

Stopped

We cannot process data once the Epic server has been stopped.

Bitmap

Internal Roaring Bitmap error

SyncError(String)

Error during chain sync

Tuple Fields of SyncError

0: String

InvalidSortAlgo

comment here

ThereIsNotPolicy

comment here

PolicyIsNotAllowed

InvalidSeed