Struct epic_chain::store::Batch

pub struct Batch<'a> { /* fields omitted */ }

An atomic batch in which all changes can be committed all at once or discarded on error.

Implementations

impl<'a> Batch<'a>

pub fn head(&self) -> Result<Tip, Error>

The head.

pub fn tail(&self) -> Result<Tip, Error>

The tail.

pub fn header_head(&self) -> Result<Tip, Error>

The current header head (may differ from chain head).

pub fn get_sync_head(&self) -> Result<Tip, Error>

Get “sync” head.

pub fn head_header(&self) -> Result<BlockHeader, Error>

Header of the block at the head of the block chain (not the same thing as header_head).

pub fn save_body_head(&self, t: &Tip) -> Result<(), Error>

Save body head to db.

pub fn save_body_tail(&self, t: &Tip) -> Result<(), Error>

Save body “tail” to db.

pub fn save_header_head(&self, t: &Tip) -> Result<(), Error>

Save header head to db.

pub fn reset_sync_head(&self) -> Result<(), Error>

Reset sync_head to the current head of the header chain.

pub fn reset_header_head(&self) -> Result<(), Error>

Reset header_head to the current head of the body chain.

pub fn save_sync_head(&self, t: &Tip) -> Result<(), Error>

Save “sync” head to db.

pub fn get_block(&self, h: &Hash) -> Result<Block, Error>

get block

pub fn block_exists(&self, h: &Hash) -> Result<bool, Error>

Does the block exist?

pub fn save_block(&self, b: &Block) -> Result<(), Error>

Save the block to the db. Note: the block header is not saved to the db here, assumes this has already been done.

pub fn save_spent_index(
&self,
h: &Hash,
spent: &Vec<CommitPos>
) -> Result<(), Error>

We maintain a “spent” index for each full block to allow the output_pos to be easily reverted during rewind.

pub fn migrate_block(
&self,
b: &Block,
version: ProtocolVersion
) -> Result<(), Error>

Migrate a block stored in the db by serializing it using the provided protocol version. Block may have been read using a previous protocol version but we do not actually care.

pub fn delete(&self, key: &[u8]) -> Result<(), Error>

Low level function to delete directly by raw key.

pub fn delete_block(&self, bh: &Hash) -> Result<(), Error>

Delete a full block. Does not delete any record associated with a block header.

pub fn save_block_header(&self, header: &BlockHeader) -> Result<(), Error>

Save block header to db.

pub fn save_output_pos_height(
&self,
commit: &Commitment,
pos: u64,
height: u64
) -> Result<(), Error>

Save output_pos and block height to index.

pub fn delete_output_pos_height(&self, commit: &Commitment) -> Result<(), Error>

Delete the output_pos index entry for a spent output.

pub fn is_match_output_pos_key(&self, key: &[u8], commit: &Commitment) -> bool

When using the output_pos iterator we have access to the index keys but not the original commitment that the key is constructed from. So we need a way of comparing a key with another commitment without reconstructing the commitment from the key bytes.

pub fn output_pos_iter(&self) -> Result<SerIterator<(u64, u64)>, Error>

Iterator over the output_pos index.

pub fn get_output_pos(&self, commit: &Commitment) -> Result<u64, Error>

Get output_pos from index.

pub fn get_output_pos_height(
&self,
commit: &Commitment
) -> Result<(u64, u64), Error>

Get output_pos and block height from index.

pub fn get_previous_header(
&self,
header: &BlockHeader
) -> Result<BlockHeader, Error>

Get the previous header.

pub fn get_block_header(&self, h: &Hash) -> Result<BlockHeader, Error>

Get block header.

pub fn save_block_sums(&self, h: &Hash, sums: &BlockSums) -> Result<(), Error>

Save block_sums for the block.

pub fn get_block_sums(&self, h: &Hash) -> Result<BlockSums, Error>

Get block_sums for the block.

pub fn get_block_input_bitmap(&self, bh: &Hash) -> Result<Bitmap, Error>

Get the block input bitmap based on our spent index. Fallback to legacy block input bitmap from the db.

pub fn get_spent_index(&self, bh: &Hash) -> Result<Vec<CommitPos>, Error>

Get the “spent index” from the db for the specified block. If we need to rewind a block then we use this to “unspend” the spent outputs.

pub fn commit(self) -> Result<(), Error>

Commits this batch. If it’s a child batch, it will be merged with the parent, otherwise the batch is written to db.

pub fn child(&mut self) -> Result<Batch<'_>, Error>

Creates a child of this batch. It will be merged with its parent on commit, abandoned otherwise.

pub fn blocks_iter(&self) -> Result<SerIterator<Block>, Error>

An iterator to all block in db