Trait epic_chain::types::TxHashsetWriteStatus

pub trait TxHashsetWriteStatus {
    fn on_setup(&self);

    fn on_validation_kernels(&self, kernels: u64, kernel_total: u64);

    fn on_validation_rproofs(&self, rproofs: u64, rproof_total: u64);

    fn on_save(&self);

    fn on_done(&self);
}

Inform the caller of the current status of a txhashset write operation, as it can take quite a while to process. Each function is called in the order defined below and can be used to provide some feedback to the caller. Functions taking arguments can be called repeatedly to update those values as the processing progresses.

Required methods

fn on_setup(&self)

First setup of the txhashset

fn on_validation_kernels(&self, kernels: u64, kernel_total: u64)

Starting kernel validation

fn on_validation_rproofs(&self, rproofs: u64, rproof_total: u64)

Starting rproof validation

fn on_save(&self)

Starting to save the txhashset and related data

fn on_done(&self)

Done writing a new txhashset